Moving the sequence optimization functions to a separate file.
