+++ codon_tools/util/__init__.py
@@ -1,2 +1,4 @@
 from .seq import *
 from .logging import logging, log_levels
+
+import codon_tools.util.seq_opt
